refactor(multisig): rename shielded presets by capability
The V1/V2/V3 suffixes implied versions of one contract, but they are
three distinct designs. Rename each by what it does:

* ShieldedMultiSig   -> ShieldedProposalMultisig    (on-chain proposal lifecycle)
* ShieldedMultiSigV2 -> ShieldedStatelessMultisig   (off-chain signatures, stateless treasury)
* ShieldedMultiSigV3 -> NativeShieldedTokenMultisig (threshold mint/burn native token)

Renames the .compact sources in presets/ along with their simulators,
witnesses, and test files, and updates every identifier, artifact path,
and the CHANGELOG entry. No behaviour change.
