+++ contracts/src/multisig/test/ShieldedStatelessMultisig.test.ts
@@ -1,10 +1,16 @@
 import { beforeEach, describe, expect, it } from 'vitest';
-import { ShieldedMultiSigV2Simulator } from './simulators/ShieldedMultiSigV2Simulator.js';
+import {
+  GENESIS_SHIELDED_COLORS,
+  makeShieldedCoin as makeCoin,
+} from '#test-utils/liveShielded.js';
+import { ShieldedStatelessMultisigSimulator } from './simulators/ShieldedStatelessMultisigSimulator.js';
 
 const RecipientKind = { ShieldedUser: 0, UnshieldedUser: 1, Contract: 2 };
 
 const INSTANCE_SALT = new Uint8Array(32).fill(0xaa);
-const COLOR = new Uint8Array(32).fill(1);
+// A shielded token type the deployer wallet holds on live (genesis-minted);
+// `fill(1)` would be unfunded on live. On dry the color is arbitrary.
+const COLOR = GENESIS_SHIELDED_COLORS.shieldedCoin1;
 const AMOUNT = 1000n;
 
 const PK1 = new Uint8Array(64).fill(0x11);
@@ -12,20 +18,24 @@ const PK2 = new Uint8Array(64).fill(0x22);
 const PK3 = new Uint8Array(64).fill(0x33);
 const NON_SIGNER_PK = new Uint8Array(64).fill(0x99);
 
-const COMMITMENT1 = ShieldedMultiSigV2Simulator.calculateSignerId(
+const COMMITMENT1 = ShieldedStatelessMultisigSimulator.calculateSignerId(
   PK1,
   INSTANCE_SALT,
 );
-const COMMITMENT2 = ShieldedMultiSigV2Simulator.calculateSignerId(
+const COMMITMENT2 = ShieldedStatelessMultisigSimulator.calculateSignerId(
   PK2,
   INSTANCE_SALT,
 );
-const COMMITMENT3 = ShieldedMultiSigV2Simulator.calculateSignerId(
+const COMMITMENT3 = ShieldedStatelessMultisigSimulator.calculateSignerId(
   PK3,
   INSTANCE_SALT,
 );
 const SIGNER_COMMITMENTS = [COMMITMENT1, COMMITMENT2, COMMITMENT3];
 
+// ECDSA verification is stubbed in the contract (`stubVerifySignature` returns
+// true), so any 64-byte value passes. Authorization is enforced only by
+// signer-commitment membership and duplicate detection — both caller-agnostic,
+// so this spec runs unchanged on live (no `ownPublicKey`-based caller identity).
 const DUMMY_SIG = new Uint8Array(64).fill(0xff);
 
 function makeRecipient(address: Uint8Array): {
@@ -33,18 +43,6 @@ function makeRecipient(address: Uint8Array): {
   address: Uint8Array;
 } {
   return { kind: RecipientKind.ShieldedUser, address };
-}
-
-function makeCoin(
-  color: Uint8Array,
-  value: bigint,
-  nonce?: Uint8Array,
-): { nonce: Uint8Array; color: Uint8Array; value: bigint } {
-  return {
-    nonce: nonce ?? new Uint8Array(32).fill(0),
-    color,
-    value,
-  };
 }
 
 function makeQualifiedCoin(
@@ -66,12 +64,12 @@ function makeQualifiedCoin(
   };
 }
 
-let multisig: ShieldedMultiSigV2Simulator;
+let multisig: ShieldedStatelessMultisigSimulator;
 
-describe('ShieldedMultiSigV2', () => {
+describe('ShieldedStatelessMultisig', () => {
   describe('constructor', () => {
     it('should initialize with 2-of-3 threshold', async () => {
-      multisig = await ShieldedMultiSigV2Simulator.create(
+      multisig = await ShieldedStatelessMultisigSimulator.create(
         INSTANCE_SALT,
         SIGNER_COMMITMENTS,
         2n,
@@ -81,7 +79,7 @@ describe('ShieldedMultiSigV2', () => {
     });
 
     it('should initialize with 1-of-3 threshold', async () => {
-      multisig = await ShieldedMultiSigV2Simulator.create(
+      multisig = await ShieldedStatelessMultisigSimulator.create(
         INSTANCE_SALT,
         SIGNER_COMMITMENTS,
         1n,
@@ -91,7 +89,7 @@ describe('ShieldedMultiSigV2', () => {
 
     it('should fail with zero threshold', async () => {
       await expect(
-        ShieldedMultiSigV2Simulator.create(
+        ShieldedStatelessMultisigSimulator.create(
           INSTANCE_SALT,
           SIGNER_COMMITMENTS,
           0n,
@@ -101,18 +99,18 @@ describe('ShieldedMultiSigV2', () => {
 
     it('should fail with threshold greater than 2', async () => {
       await expect(
-        ShieldedMultiSigV2Simulator.create(
+        ShieldedStatelessMultisigSimulator.create(
           INSTANCE_SALT,
           SIGNER_COMMITMENTS,
           3n,
         ),
       ).rejects.toThrow(
-        'ShieldedMultiSigV2: threshold cannot exceed 2 (execute verifies at most 2 signatures)',
+        'ShieldedStatelessMultisig: threshold cannot exceed 2 (execute verifies at most 2 signatures)',
       );
     });
 
     it('should register all signer commitments', async () => {
-      multisig = await ShieldedMultiSigV2Simulator.create(
+      multisig = await ShieldedStatelessMultisigSimulator.create(
         INSTANCE_SALT,
         SIGNER_COMMITMENTS,
         2n,
@@ -123,12 +121,12 @@ describe('ShieldedMultiSigV2', () => {
     });
 
     it('should reject a non-signer commitment', async () => {
-      multisig = await ShieldedMultiSigV2Simulator.create(
+      multisig = await ShieldedStatelessMultisigSimulator.create(
         INSTANCE_SALT,
         SIGNER_COMMITMENTS,
         2n,
       );
-      const unknown = ShieldedMultiSigV2Simulator.calculateSignerId(
+      const unknown = ShieldedStatelessMultisigSimulator.calculateSignerId(
         NON_SIGNER_PK,
         INSTANCE_SALT,
       );
@@ -138,7 +136,7 @@ describe('ShieldedMultiSigV2', () => {
 
   describe('when initialized', () => {
     beforeEach(async () => {
-      multisig = await ShieldedMultiSigV2Simulator.create(
+      multisig = await ShieldedStatelessMultisigSimulator.create(
         INSTANCE_SALT,
         SIGNER_COMMITMENTS,
         2n,
