+++ contracts/src/multisig/test/witnesses/NativeShieldedTokenMultisigWitnesses.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (multisig/witnesses/NativeShieldedTokenMultisigWitnesses.ts)
+
+export type NativeShieldedTokenMultisigPrivateState = Record<string, never>;
+export const NativeShieldedTokenMultisigPrivateState: NativeShieldedTokenMultisigPrivateState =
+  {};
+export const NativeShieldedTokenMultisigWitnesses = () => ({});
