--- contracts/src/multisig/test/witnesses/ShieldedMultiSigV2Witnesses.ts
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.3.0-alpha (multisig/witnesses/ShieldedMultiSigV2Witnesses.ts)
-
-export type ShieldedMultiSigV2PrivateState = Record<string, never>;
-export const ShieldedMultiSigV2PrivateState: ShieldedMultiSigV2PrivateState =
-  {};
-export const ShieldedMultiSigV2Witnesses = () => ({});
