--- contracts/src/multisig/test/witnesses/ShieldedMultiSigV3Witnesses.ts
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (multisig/witnesses/ShieldedMultiSigV3Witnesses.ts)
-
-export type ShieldedMultiSigV3PrivateState = Record<string, never>;
-export const ShieldedMultiSigV3PrivateState: ShieldedMultiSigV3PrivateState =
-  {};
-export const ShieldedMultiSigV3Witnesses = () => ({});
