--- contracts/src/multisig/test/witnesses/ShieldedMultiSigWitnesses.ts
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.3.0-alpha (multisig/witnesses/ShieldedMultiSigWitnesses.ts)
-
-export type ShieldedMultiSigPrivateState = Record<string, never>;
-export const ShieldedMultiSigPrivateState: ShieldedMultiSigPrivateState = {};
-export const ShieldedMultiSigWitnesses = () => ({});
