+++ contracts/src/multisig/test/witnesses/ShieldedProposalMultisigWitnesses.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+// OpenZeppelin Compact Contracts v0.3.0-alpha (multisig/witnesses/ShieldedProposalMultisigWitnesses.ts)
+
+export type ShieldedProposalMultisigPrivateState = Record<string, never>;
+export const ShieldedProposalMultisigPrivateState: ShieldedProposalMultisigPrivateState =
+  {};
+export const ShieldedProposalMultisigWitnesses = () => ({});
