+++ contracts/src/multisig/test/witnesses/ShieldedStatelessMultisigWitnesses.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+// OpenZeppelin Compact Contracts v0.3.0-alpha (multisig/witnesses/ShieldedStatelessMultisigWitnesses.ts)
+
+export type ShieldedStatelessMultisigPrivateState = Record<string, never>;
+export const ShieldedStatelessMultisigPrivateState: ShieldedStatelessMultisigPrivateState =
+  {};
+export const ShieldedStatelessMultisigWitnesses = () => ({});
